feat(refactor): break down page into components
this commit tidies things up by moving the hero, loan products, and loader sections into their own components.

this makes the main page much cleaner and the overall structure easier to follow. each piece now lives in its own file, which should make future updates a breeze.
